--- council/runners/budget.py
@@ -1,2 +1,0 @@
-# deprecated
-from council.contexts import Budget, Consumption, InfiniteBudget
