+++ council/runners/types.py
@@ -1,7 +1,6 @@
 from typing import Callable, Iterable, Any
 
-from council.contexts import ChainContext
-from .budget import Budget
+from council.contexts import Budget, ChainContext
 
 RunnerPredicate = Callable[[ChainContext, Budget], bool]
 RunnerGenerator = Callable[[ChainContext, Budget], Iterable[Any]]
