+++ council/skills/skill_base.py
@@ -4,8 +4,8 @@ import logging
 from typing import Any
 from abc import abstractmethod
 
-from council.contexts import SkillContext, ChatMessage
-from council.runners import SkillRunnerBase, Budget
+from council.contexts import Budget, SkillContext, ChatMessage
+from council.runners import SkillRunnerBase
 
 logger = logging.getLogger(__name__)
