+++ tests/integration/evaluators/test_llm_evaluator.py
@@ -2,11 +2,10 @@ import unittest
 
 import dotenv
 
-from council.contexts import AgentContext, ChatMessage, ChainContext
+from council.contexts import AgentContext, Budget, ChatMessage, ChainContext
 from council.llm import AzureLLM
 from council.evaluators import LLMEvaluator
 from council.mocks import MockMonitored
-from council.runners import Budget
 
 
 class TestLlmEvaluator(unittest.TestCase):
