+++ tests/integration/skill/test_google_skills.py
@@ -2,8 +2,7 @@ import unittest
 
 import dotenv
 
-from council.contexts import ChainContext
-from council.runners import Budget
+from council.contexts import ChainContext, Budget
 from council.skills.google import GoogleNewsSkill, GoogleSearchSkill
 from council.skills.google.google_context import GoogleNewsSearchEngine, GoogleSearchEngine
 from council.utils.option import OptionException
