+++ tests/integration/skill/test_llm_skill.py
@@ -5,10 +5,9 @@ from typing import List
 import dotenv
 
 from council.agents import Agent
-from council.contexts import AgentContext
+from council.contexts import AgentContext, Budget
 from council.llm import AzureLLM, LLMMessage
 from council.mocks import MockLLM
-from council.runners import Budget
 from council.skills.llm_skill import LLMSkill
